+++ client/assets/javascripts/actions/FluxUserActions.js
@@ -1,12 +1,14 @@
 import alt from '../FluxAlt';
 import _ from 'lodash';
 import UserAPI from '../utils/api/UserAPI';
+import FluxCollectionActions from './FluxCollectionActions'
 
 class FluxUserActions {
   fetchData(id) {
     UserAPI.getUser(id,
       (data) => {
         this.actions.updateData(data);
+        FluxCollectionActions.fetchedCollections(data.collections);
       },
       (error) => {
         this.actions.registerError(error);