+++ client/assets/javascripts/components/collections/CollectionModal.jsx
@@ -3,6 +3,7 @@ import _ from 'lodash';
 import { Link, Navigation } from 'react-router';
 import CollectionBox from './CollectionBox';
 import Modal from 'react-modal';
+import CollectionStore from '../../stores/CollectionStore';
 import FluxAlertActions from '../../actions/FluxAlertActions';
 import FluxCollectionActions from '../../actions/FluxCollectionActions';
 import FluxNotificationsActions from '../../actions/FluxNotificationsActions'
@@ -20,12 +21,15 @@ const CollectionMixin = {
 
   renderCollectionModal: function() {
     return (
-      <CollectionModal
-        ref='collectionModal'
-        visible={this.collectionModalVisible()}
-        show={this.showCollectionModal}
-        close={this.closeCollectionModal}
-        onSaveCollection={this.onSaveCollection} />
+      <div>
+        <CollectionModal
+          ref='collectionModal'
+          visible={this.collectionModalVisible()}
+          show={this.showCollectionModal}
+          close={this.closeCollectionModal}
+          onSaveCollection={this.onSaveCollection} />
+        {this.renderCollectionShareModal()}
+      </div>
     )
   },
 
@@ -34,34 +38,69 @@ const CollectionMixin = {
   },
 
   closeCollectionModal: function() {
+    $('body').removeClass('no-scroll');
     this.setState({showCollectionModal: false})
-    this.refs.collectionModal.clearCollection()
+    if(!(this.props.route && this.props.route.name == "CollectionPage")) {
+      FluxCollectionActions.clearCollection();
+    }
   },
 
   showCollectionModal: function() {
+    $(window).scrollTop(0);
+    $('body').addClass('no-scroll');
     this.setState({showCollectionModal: true})
   },
 
-  showCollectionModalForEditing: function(collection) {
-    this.refs.collectionModal.setCollection(collection)
+  showCollectionModalWithProduct: function(product) {
+    FluxCollectionActions.fetchedCollection({
+      title: '', description: '', products: [product]
+    });
     this.showCollectionModal()
   },
 
+  showCollectionModalForEditing: function(collection) {
+    FluxCollectionActions.fetchedCollection(collection);
+    this.showCollectionModal()
+  },
+
+  // TODO: Refactor this function, it's way too complex.
   onSaveCollection: function(collection, resolve) {
     let _this = this;
+
+    // Checks whether we're on a page and whether the collection that was just updated
+    // contains the passed in collection. If it doesn't, remove it.
+    let checkAndRemoveCollectionFromStore = function(collection) {
+      // Check whether we're on the product page
+      if(_this.context.router.state.components[0].displayName == 'ProductPage') {
+        // Get a list of products still in the collection
+        let productIDs = _.map(collection.products, function(product) {return product.id})
+        // If this product isn't one of them, remove the collection from the collections store.
+        if(productIDs.indexOf(_this.props.data.id) == -1) {
+          FluxCollectionActions.removeCollection(collection.id);
+        }
+      }
+    }
+
+    // If we have a collection id, then the collection is persisted on the backend and just needs updating.
     if(collection.id) {
-      FluxCollectionActions.updateCollection(collection.id, collection, resolve)
+      FluxCollectionActions.updateCollection(collection.id, collection, function(data) {
+        resolve();
+
+        checkAndRemoveCollectionFromStore(data);
+      })
+    // Otherwise it's a new collection and we just need to save.
     } else {
       FluxCollectionActions.createCollection(collection, function(collection) {
         resolve();
 
+        // Show an alert after saving that asks the user whether they want to share the collection or not.
         FluxAlertActions.showAlert({
           title: 'Your Collection was successfully created!',
           message: 'You can privately share this Collection with other users in Fletcher, or make it available to everyone by making it public.',
           success: 'Share',
           cancel: 'Not now',
           successCallback: function() {
-            _this.showCollectionShareModalForEditing(collection)
+            _this.showCollectionShareModal(collection)
           }
         })
       })
@@ -74,25 +113,22 @@ const CollectionModal = React.createClass ({
 
   getInitialState: function() {
     return {
-      collection: {
-        title: '',
-        description: '',
-        privacy: 'hidden',
-        products: []
+      data: {
+        collection: {
+          title: '',
+          description: '',
+          products: []
+        }
       }
     }
   },
 
-  setCollection: function(collection) {
-    this.setState({
-      collection: {
-        id: collection.id,
-        title: collection.title,
-        description: collection.description,
-        privacy: collection.privacy,
-        products: collection.products
-      }
-    })
+  componentDidMount: function() {
+    CollectionStore.listen(this.onChange);
+  },
+
+  onChange: function(data) {
+    this.setState(data);
   },
 
   clearCollection: function() {
@@ -112,22 +148,25 @@ const CollectionModal = React.createClass ({
   },
 
   getProductIDs: function() {
-    return _.map(this.state.collection.products, function(product) {
+    return _.map(this.state.data.collection.products, function(product) {
       return product.id
     })
   },
 
   removeProduct: function(product_id) {
-    let products = this.state.collection.products.filter(function(product) {
-        return product.id !== product_id;
+    let products = this.state.data.collection.products.filter(function(product) {
+      return product.id !== product_id;
     });
 
-    this.setState({product_name: null, collection: {products: products}})
+    let data = this.state.data
+    data.collection.products = products
+
+    this.setState({product_name: null, data: data})
   },
 
   addProduct: function(product, selected) {
     if(selected) {
-      let newProducts = this.state.collection.products
+      let newProducts = this.state.data.collection.products
       newProducts.push(product)
       this.setState({product_name: null, collection: {products: newProducts}})
     } else {
@@ -140,9 +179,9 @@ const CollectionModal = React.createClass ({
     let _this = this
 
     let collection = {
-      id: this.state.collection.id,
-      title: this.state.collection.title,
-      description: this.state.collection.description,
+      id: this.state.data.collection.id,
+      title: this.state.data.collection.title,
+      description: this.state.data.collection.description,
       privacy: e.currentTarget.dataset.privacy,
       products: this.getProductIDs()
     }
@@ -162,9 +201,13 @@ const CollectionModal = React.createClass ({
   },
 
   onChangeField: function(name, e) {
-    let hash = {}
+    let hash = this.state.data.collection;
     hash[name] = e.currentTarget.value
-    this.setState(hash)
+    this.setState({
+      data: {
+        collection: hash
+      }
+    })
   },
 
   renderTextFields: function() {
@@ -175,7 +218,7 @@ const CollectionModal = React.createClass ({
                 placeholder='Title'
                 name='collection[title]'
                 ref='collection_title'
-                value={this.state.collection.title}
+                value={this.state.data.collection.title}
                 onFocus={this.onFocus}
                 onBlur={this.onBlur}
                 onChange={(e) => this.onChangeField('title', e)} />
@@ -185,7 +228,7 @@ const CollectionModal = React.createClass ({
                   name='collection[description]'
                   rows='10'
                   ref='collection_description'
-                  value={this.state.collection.description}
+                  value={this.state.data.collection.description}
                   onFocus={this.onFocus}
                   onBlur={this.onBlur}
                   onChange={(e) => this.onChangeField('description', e)} />
@@ -199,7 +242,9 @@ const CollectionModal = React.createClass ({
                     value={this.state.product_name}
                     helpMessage={'Add Product'}
                     hideLabel={true}
-                    onSetProduct={this.addProduct} />
+                    onSetProduct={this.addProduct}
+                    placeholder={'Search products and add them'}
+                    noEmptySubmit={true} />
     )
   },
 
@@ -208,18 +253,18 @@ const CollectionModal = React.createClass ({
       <Results
         type='collection-product'
         onRemove={this.removeProduct}
-        data={{data: this.state.collection.products}} />
+        data={{data: this.state.data.collection.products}} />
     )
   },
 
   buttonsDisabled: function() {
-    return !(this.state.collection.title.length > 0 &&
-             this.state.collection.description.length > 0 &&
-             this.state.collection.products.length > 0)
+    return !(this.state.data.collection.title.length > 0 &&
+             this.state.data.collection.description.length > 0 &&
+             this.state.data.collection.products.length > 0)
   },
 
   updating: function() {
-    return !!this.state.collection.id
+    return !!this.state.data.collection.id
   },
 
   renderSubmissionButtons: function() {
@@ -229,7 +274,7 @@ const CollectionModal = React.createClass ({
         <div className='buttons'>
           <button className='btn btn-red btn-round'
                   onClick={this.submitForm}
-                  data-privacy={this.state.collection.privacy}
+                  data-privacy={this.state.data.collection.privacy}
                   disabled={disabled}>Update</button>
         </div>
       )
@@ -248,7 +293,7 @@ const CollectionModal = React.createClass ({
   renderCollectionForm: function() {
     return (
       <div className='row'>
-        <form className='col-xs-10 col-xs-offset-1 form collection'
+        <form className='col-xs-12 col-sm-10 col-sm-offset-1 form collection'
               ref='collection_form'>
           {this.renderTextFields()}
           {this.renderProductTypeahead()}
@@ -263,6 +308,7 @@ const CollectionModal = React.createClass ({
     return (
       <Modal
         isOpen={this.props.visible}>
+        <div className='back-button' onClick={this.close}>{"< Close"}</div>
         <div className='header'>
           <span className='title'>
             {this.updating() ? 'Update' : 'Create'} Collection