+++ client/assets/javascripts/components/collections/CollectionPage.jsx
@@ -2,23 +2,32 @@ import _ from 'lodash';
 import React from 'react';
 import timeago from 'timeago';
 import { Link, Navigation } from 'react-router';
-import FluxCollectionActions from '../../actions/FluxCollectionActions'
+import FluxCollectionActions from '../../actions/FluxCollectionActions';
+import FluxAlertActions from '../../actions/FluxAlertActions';
+import FluxNotificationsActions from '../../actions/FluxNotificationsActions';
 import CollectionStore from '../../stores/CollectionStore'
 import Results from '../search/Results'
+import { CollectionMixin } from './CollectionModal';
 
 const CollectionPage = React.createClass({
   displayName: 'CollectionPage',
-  mixins: [ Navigation ],
+  mixins: [ Navigation, CollectionMixin ],
+
+  contextTypes: {
+    currentUser: React.PropTypes.object.isRequired
+  },
 
   getInitialState: function() {
     return {
       data: {
-        title: '',
-        description: '',
-        user: {
-          name: ''
-        },
-        products: []
+        collection: {
+          title: '',
+          description: '',
+          user: {
+            name: ''
+          },
+          products: []
+        }
       }
     };
   },
@@ -46,18 +55,64 @@ const CollectionPage = React.createClass({
     this.setState(data);
   },
 
+  ownedByUser: function() {
+    return this.context.currentUser.id == this.state.data.collection.user.id
+  },
+
+  deleteCollection: function() {
+    let _this = this;
+    let collection_name = this.state.data.collection.title;
+
+    FluxAlertActions.showAlert({
+      title: `Are you sure you want to delete ${collection_name}`,
+      success: 'Yes, delete it',
+      cancel: "No, don't delete it",
+      successCallback: function() {
+        FluxCollectionActions.deleteCollection({
+          id: _this.props.params.id,
+          name: collection_name
+        })
+        _this.context.router.transitionTo('/app')
+      }
+    })
+  },
+
+  renderEditButtons: function () {
+    if(this.ownedByUser()) {
+      return (
+        <div className='user-buttons'>
+          <div  className='btn btn-red-inverted btn-round'
+                onClick={() => this.showCollectionModalForEditing(this.state.data.collection)}>
+            Edit
+          </div>
+          <div  className='btn btn-red-inverted btn-round'
+                onClick={() => this.showCollectionShareModal(this.state.data.collection)}>
+            Share
+          </div>
+          <div  className='btn btn-red-inverted btn-round'
+                onClick={this.deleteCollection}>
+            Delete
+          </div>
+        </div>
+      )
+    }
+  },
+
   renderCollectionInfo: function () {
     return (
       <div className='col-xs-12 tag-header'>
         <div className='title'>
-          { this.state.data.title }
+          { this.state.data.collection.title }
         </div>
         <div className='created_by'>
-          Created by { this.state.data.user.name }
+          Created by <Link to={`/app/users/${this.state.data.collection.user.id}`}>
+            { this.state.data.collection.user.name }
+          </Link>
         </div>
         <div className='description'>
-          { this.state.data.description }
+          { this.state.data.collection.description }
         </div>
+        {this.renderEditButtons()}
       </div>
     )
   },
@@ -67,7 +122,7 @@ const CollectionPage = React.createClass({
       <div className='col-xs-12 results-container'>
         <Results
           type='products'
-          data={{total: this.state.data.products.length, data: this.state.data.products}}
+          data={{total: this.state.data.collection.products.length, data: this.state.data.collection.products}}
           showImages={true}
           topLeft='count' />
       </div>
@@ -81,6 +136,9 @@ const CollectionPage = React.createClass({
           { this.renderCollectionInfo() }
           { this.renderResults() }
         </div>
+
+        {this.renderCollectionModal()}
+        {this.renderCollectionShareModal()}
       </div>
     );
   }