+++ client/assets/javascripts/components/products/ProductPageDesktopVersion.jsx
@@ -12,7 +12,7 @@ import Tags from '../Tags';
 import UrlHelper from '../../utils/helpers/UrlHelper'
 import FileHelper from '../../utils/helpers/FileHelper'
 import RelatedProducts from './RelatedProducts'
-import CollectionsCollection from '../collections/CollectionsCollection'
+import CollectionsCollection from '../collections/CollectionsCollection';
 
 const ProductPageDesktopVersion = React.createClass({
   displayName: 'ProductPageDesktopVersion',
@@ -214,12 +214,18 @@ const ProductPageDesktopVersion = React.createClass({
           </div>
           <div className='col-xs-9'>
             <Reviews productID={this.id()} ref='reviews' />
-            <div className='collections-container' ref='collections'>
+            <div className='collections-container hide' ref='collections'>
               <div className='header'>
               </div>
+              <div className='placeholder-section message'>
+                Collections are created by users to group products they are interested. They can even be shared or made public. Create one yourself!
+              </div>
+              <div className='btn btn-round btn-red' onClick={() => this.props.showAddToCollectionModal(this.props.data)}>
+                Add product to a Collection
+              </div>
               <CollectionsCollection
-                className='hide'
-                collections={this.props.data.collections}/>
+                onEdit={this.props.onCollectionEdit}
+                onShare={this.props.onCollectionShare} />
             </div>
             <div className='placeholder-section hide' ref='custom'>
               Feature Coming Soon